add routing to register
