+++ src/Register.js
@@ -1,0 +1,15 @@
+import React from "react";
+import Header from "./components/Header";
+
+function Register() {
+
+  return (
+    <div>
+      <Header/>
+      <hr />
+      
+    </div>
+  );
+}
+
+export default Register;
