+++ src/components/Header.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+function Header() {
+  return (
+    <div class="LoginAndRegisterHeader">
+      <img
+        src="https://firebasestorage.googleapis.com/v0/b/budgify-ed7a9.appspot.com/o/LogoBlack.png?alt=media&token=203a5607-2e84-47bd-8417-49e7c5c93fd5"
+        id="headerImage" alt=""/>
+    </div>
+  );
+}
+
+export default Header;
