+++ src/styles.css
@@ -2,10 +2,11 @@
 @import url(https://fonts.googleapis.com/css?family=Roboto+Condensed:300,300italic,400,400italic,700,700italic&subset=latin,latin-ext,cyrillic-ext,cyrillic,greek-ext,greek,vietnamese);
 @import url(https://fonts.googleapis.com/css?family=Roboto+Slab:400,100,300,700&subset=latin,latin-ext,greek-ext,greek,vietnamese,cyrillic,cyrillic-ext);
 body {
-  background: linear-gradient(#646bc3, #c15fef, black);
-  background-size: cover;
 }
 .container {
+  
+  background: linear-gradient(#646bc3, #c15fef, black);
+  background-size: cover;
   height: 100vh;
   margin-left: -5%;
   margin-right: -5%;
@@ -115,4 +116,15 @@ footer {
   font-family: "Roboto", sans-serif;
   font-weight: 120;
   color: #fff;
+}
+
+.LoginAndRegisterHeader{
+  margin: 0px;
+  border: 0px;
+  text-align: center;
+}
+
+#headerImage{
+  width: 50px;
+  opacity: 0.8;
 }
